+++ src/main/java/com/ferreteriafc/api/backend/persistence/entity/Category.java
@@ -1,13 +1,13 @@
 package com.ferreteriafc.api.backend.persistence.entity;
 
+import java.util.ArrayList;
+import java.util.List;
+
 import jakarta.persistence.*;
 
 import lombok.*;
 
 import com.ferreteriafc.api.backend.domain.utils.Constant;
-
-import java.util.ArrayList;
-import java.util.List;
 
 @Entity
 @Table(name = "categories")
@@ -28,7 +28,7 @@ public class Category {
     @OneToMany(mappedBy = "category", fetch = FetchType.LAZY)
     private List<Product> products;
 
-    public Category(Long id, String name, String imageUrl) {
+    public Category(Integer id, String name, String imageUrl) {
         this.products = new ArrayList<>();
         this.id = id;
         this.name = name;
